bug fix: invalid float*string multiplication

+++ instapy/util.py
@@ -924,9 +924,8 @@ def highlight_print(username=None, message=None, priority=None, level=None, logg
         upper_char = "~"
         lower_char = None
 
-
     if show_logs == True:
-        print("\n{}".format(upper_char * ceil(output_len/len(upper_char))))
+        print("\n{}".format(upper_char * int(ceil(output_len/len(upper_char)))))
 
     if level == "info":
         logger.info(message)
